fix(ui): apply all review changes for weight configuration #227
