+++ frontend/src/App.jsx
@@ -31,7 +31,8 @@ import './styles.css';
 import AdvisorRequestsPage from './AdvisorRequestsPage';
 import ProfessorCommitteeSubmissionsPage from './ProfessorCommitteeSubmissionsPage';
 import CommitteeGradingPage from './CommitteeGradingPage';
-import WeightConfiguration from './components/WeightConfiguration';
+import CoordinatorWeightConfigurationPage from './CoordinatorWeightConfigurationPage';
+import AuthGuard from './components/AuthGuard';
 
 export default function App() {
   return (
@@ -67,7 +68,7 @@ export default function App() {
               <Route path="/coordinator/groups/manage" element={<CoordinatorGroupMembershipPage />} />
               <Route path="/coordinator/groups/transfer" element={<CoordinatorAdvisorTransferPage />} />
               <Route path="/coordinator/groups/cleanup" element={<GroupCleanupPage role="COORDINATOR" />} />
-              <Route path="/coordinator/grading/weight-configuration" element={<WeightConfiguration />} />
+              <Route path="/coordinator/grading/weight-configuration" element={<AuthGuard allowedRoles={['COORDINATOR']}><CoordinatorWeightConfigurationPage /></AuthGuard>} />
               <Route path="/groups/:groupId" element={<GroupPage />} />
               <Route path="/advisor/requests" element={<AdvisorRequestsPage />} />
               <Route path="/professors/committee-submissions" element={<ProfessorCommitteeSubmissionsPage />} />
