+++ frontend/src/components/AppShell.jsx
@@ -36,6 +36,7 @@ const roleMenuSections = {
         { to: '/coordinator/groups/manage', label: 'Group Membership Edit', icon: 'GM' },
         { to: '/coordinator/groups/transfer', label: 'Advisor Transfer', icon: 'AT' },
         { to: '/coordinator/groups/cleanup', label: 'Group Cleanup', icon: 'GC' },
+        { to: '/coordinator/grading/weight-configuration', label: 'Weight Configuration', icon: 'WC' },
       ],
     },
   ],
